experiments/ai_model_exploration/chatterbox_tts/requirements.txt
click>=8.1.0
rich>=13.7.0
requests>=2.31.0
python-dotenv>=1.0.0
